+++ src/App.tsx
@@ -1,8 +1,9 @@
 import React from 'react';
 import { ConfigProvider, App as AntApp, theme } from 'antd';
 import { Provider } from 'react-redux';
+import { PersistGate } from 'redux-persist/integration/react';
 import { ThemeProvider } from 'styled-components';
-import { store } from './store/store';
+import { store, persistor } from './store/store';
 import { MainApp } from './pages/MainApp';
 import { GlobalStyle } from './styles/GlobalStyle';
 import { theme as appTheme } from './styles/theme';
@@ -16,14 +17,16 @@ const App: React.FC = () => {
 
   return (
     <Provider store={store}>
-      <ConfigProvider theme={{ algorithm: theme.darkAlgorithm }}>
-        <AntApp>
-          <ThemeProvider theme={appTheme}>
-            <GlobalStyle />
-            <MainApp onError={handleError} />
-          </ThemeProvider>
-        </AntApp>
-      </ConfigProvider>
+      <PersistGate loading={null} persistor={persistor}>
+        <ConfigProvider theme={{ algorithm: theme.darkAlgorithm }}>
+          <AntApp>
+            <ThemeProvider theme={appTheme}>
+              <GlobalStyle />
+              <MainApp onError={handleError} />
+            </ThemeProvider>
+          </AntApp>
+        </ConfigProvider>
+      </PersistGate>
     </Provider>
   );
 };